+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Method to subscribe to emergency messages.
+- Method to subscribe to register value updates.
 - Add is_node_id_dependent property to CanopenRegister
 
 ### Fixed